requirements.txt
apache-airflow==2.8.0
dbt-postgres==1.7.0
pandas==2.2.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
loguru==0.7.2
pytest==7.4.4
ruff==0.1.14
